Unify webpack builds, create simple Game renderer
I prefer (for now) one longer webpack config than two, so I merged them.

Building for either node or web required a few changes:
* plugins are all specific to the web build
* handlebars has a dependency on node (fs) by default; incorporated handlebars-loader

Fixed a missing dependency on ‘./biomes’

World navigation is not currently spatially accurate. The orientations of the faces of the dodecahedron are not the same as the ones on the die, and it is probably easier to update the map in sectors/factory than it is to reorient the renderer’s faces (those orientations are determined by the underlying geometry and UV coordinates of how the dodecahedron was generated)