game, view-face components
`game` loads and runs the game engine, and owns a reference to the dispatch callback. It updates the terminal with new text.

note: I don’t have a new terminal built yet, but you can still control the game from the browser console by running
```
const d = document.getElementById(‘terminal’).components.game.dispatch
d(‘go club’)
```

`view-face` enables an entity to automatically position itself a set distance along the normals of the target mesh. This is used for the camera as well as the terminal text.